[FLINK-35934] Add CompiledPlan annotations to BatchExecValues
