+++ flink-table/flink-table-planner/src/test/java/org/apache/flink/table/planner/plan/nodes/exec/batch/ValuesBatchRestoreTest.java
@@ -16,22 +16,24 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.planner.plan.nodes.exec.stream;
+package org.apache.flink.table.planner.plan.nodes.exec.batch;
 
-import org.apache.flink.table.test.program.SinkTestStep;
+import org.apache.flink.table.planner.plan.nodes.exec.common.ValuesTestPrograms;
+import org.apache.flink.table.planner.plan.nodes.exec.testutils.BatchRestoreTestBase;
 import org.apache.flink.table.test.program.TableTestProgram;
 
-/** {@link TableTestProgram} definitions for testing {@link StreamExecValues}. */
-public class ValuesTestPrograms {
+import java.util.Collections;
+import java.util.List;
 
-    static final TableTestProgram VALUES_TEST =
-            TableTestProgram.of("values-test", "validates values node")
-                    .setupTableSink(
-                            SinkTestStep.newBuilder("sink_t")
-                                    .addSchema("b INT", "a INT", "c VARCHAR")
-                                    .consumedValues("+I[1, 2, Hi]", "+I[3, 4, Hello]")
-                                    .build())
-                    .runSql(
-                            "INSERT INTO sink_t SELECT * FROM (VALUES (1, 2, 'Hi'), (3, 4, 'Hello'))")
-                    .build();
+/** Batch Compiled Plan tests for {@link BatchExecValues}. */
+public class ValuesBatchRestoreTest extends BatchRestoreTestBase {
+
+    public ValuesBatchRestoreTest() {
+        super(BatchExecValues.class);
+    }
+
+    @Override
+    public List<TableTestProgram> programs() {
+        return Collections.singletonList(ValuesTestPrograms.VALUES_TEST);
+    }
 }
